requirements.txt
streamlit>=1.28.0
google-cloud-storage>=2.8.0
PyPDF2>=3.0.0
python-dotenv>=1.0.0
vertexai>=0.6.0
python-docx>=0.8.11
tenacity>=8.2.3
diskcache>=5.6.3
pandas>=2.0.0
tiktoken>=0.5.1